ERP-6
Criação de autenticação do login, criação de testes para o login e criaçã ode mixins para os produtos e clientes
