+++ src/views/ControleDeClientes.vue
@@ -102,209 +102,18 @@
 </template>
 
 <script>
-import clienteService from "@/services/cliente-service";
-import Cliente from "@/models/cliente-model";
-import conversorData from "@/utils/conversor-data";
 import Button from "@/components/Button/Button.vue";
+import ClienteMixins from "../mixins/cliente-mixin";
 
 export default {
   name: "ControleDeClientes",
+  mixins: [ClienteMixins],
   components: {
     Button,
   },
-  filters: {
-    data(data) {
-      return conversorData.aplicarMascaraDataHoraEmDataIso(data);
-    },
-  },
-  data() {
-    return {
-      clientes: [],
-      pesquisa: "",
-      filtro: "nome",
-      tiposFiltro: ["nome", "cpfOuCnpj", "email"],
-      headers: [
-        {
-          text: "ID",
-          value: "id",
-        },
-        {
-          text: "Nome",
-          value: "nome",
-        },
-        {
-          text: "CPF/CNPJ",
-          value: "cpfOuCnpj",
-        },
-        {
-          text: "E-mail",
-          value: "email",
-        },
-        {
-          text: "Telefone",
-          value: "telefone",
-        },
-        {
-          text: "ID do Usuário",
-          value: "idUsuario",
-        },
-        {
-          text: "Data de Cadastro",
-          value: "dataCadastro",
-        },
-      ],
-
-      modalAberto: false,
-      cliente: new Cliente(),
-      edicao: false,
-      manterAberto: false,
-    };
-  },
-
-  created() {
-    this.obterTodosOsClientes();
-  },
-
-  computed: {
-    clientesFiltrados() {
-      if (!this.pesquisa) {
-        return this.clientes;
-      }
-      const pesquisaLowerCase = this.pesquisa.toLowerCase();
-      return this.clientes.filter((cliente) =>
-        cliente[this.filtro].toLowerCase().includes(pesquisaLowerCase)
-      );
-    },
-  },
-
   methods: {
-    abrirModalAdicao() {
-      this.cliente = new Cliente();
-      this.cliente.dataCadastro = new Date();
-      this.edicao = false;
-      this.manterAberto = false;
-      this.modalAberto = true;
-    },
-
-    abrirModalEdicao(cliente) {
-      this.cliente = new Cliente(cliente);
-      this.edicao = true;
-      this.modalAberto = true;
-    },
-
-    validarCliente(cliente) {
-      return (
-        cliente.nome &&
-        cliente.email &&
-        cliente.telefone &&
-        cliente.cpfOuCnpj &&
-        cliente.idUsuario
-      );
-    },
-
-    async adicionarCliente() {
-      if (this.validarCliente(this.cliente)) {
-        try {
-          await clienteService.cadastrar(this.cliente);
-          this.cliente = new Cliente();
-          this.obterTodosOsClientes();
-          if (!this.manterAberto) {
-            this.modalAberto = false;
-          }
-          this.$swal.fire(
-            "Sucesso",
-            "Cliente adicionado com sucesso",
-            "success"
-          );
-        } catch (error) {
-          console.error("Erro ao adicionar o cliente:", error);
-          this.$swal.fire(
-            "Erro",
-            "Ocorreu um erro ao adicionar o cliente",
-            "error"
-          );
-        }
-      } else {
-        this.$swal.fire(
-          "Erro",
-          "Por favor, preencha todos os campos obrigatórios",
-          "error"
-        );
-      }
-    },
-
     visualizarCards() {
       this.$router.push("/clientes");
-    },
-
-    async editarCliente() {
-      try {
-        await clienteService.atualizar(this.cliente);
-        this.cliente = new Cliente();
-        this.obterTodosOsClientes();
-        if (!this.manterAberto) {
-          this.modalAberto = false;
-        }
-        this.$swal.fire("Sucesso", "Cliente editado com sucesso", "success");
-      } catch (error) {
-        console.error("Erro ao editar o cliente:", error);
-        this.$swal.fire("Erro", "Ocorreu um erro ao editar o cliente", "error");
-      }
-    },
-
-    ordenarClientes(cliente, clientes) {
-      return cliente.id < clientes.id ? -1 : cliente.id > clientes.id ? 1 : 0;
-    },
-
-    async obterTodosOsClientes() {
-      try {
-        const response = await clienteService.obterTodos();
-        let clientes = response.data.map((c) => new Cliente(c));
-
-        this.clientes = clientes.sort(this.ordenarClientes).reverse();
-      } catch (error) {
-        console.log(error);
-        this.$swal.fire(
-          "Erro",
-          "Ocorreu um erro ao obter os clientes",
-          "error"
-        );
-      }
-    },
-
-    excluirCliente(id) {
-      this.$swal
-        .fire({
-          title: "Confirmar exclusão",
-          text: "Deseja excluir este cliente?",
-          icon: "warning",
-          showCancelButton: true,
-          confirmButtonText: "Sim, excluir",
-          cancelButtonText: "Cancelar",
-        })
-        .then(async (result) => {
-          if (result.isConfirmed) {
-            try {
-              await clienteService.deletar(id);
-              this.obterTodosOsClientes();
-              if (!this.manterAberto) {
-                this.modalAberto = false;
-              }
-              this.$swal.fire(
-                "Sucesso",
-                "Cliente excluído com sucesso",
-                "success"
-              );
-            } catch (error) {
-              console.error("Erro ao excluir o cliente:", error);
-              this.$swal.fire(
-                "Erro",
-                "Ocorreu um erro ao excluir o cliente",
-                "error"
-              );
-            }
-          }
-        });
     },
   },
 };
